Added code to remove the JUnit XML file for a test in the tearDown() method only if the file exists.

+++ test/Stagehand/TestRunner/Runner/TestCase.php
@@ -89,7 +89,9 @@ abstract class TestCase extends FactoryAwareTestCase
     protected function tearDown()
     {
         parent::tearDown();
-        unlink($this->junitXMLFile);
+        if (file_exists($this->junitXMLFile)) {
+            unlink($this->junitXMLFile);
+        }
     }
 
     protected function assertTestCaseCount($count)
